requirements.txt
paho-mqtt
beautifulsoup4
requests
